+++ lib/os/lua.ex
@@ -141,9 +141,11 @@ defmodule FarmbotOS.Lua do
       get_tool: &DataManipulation.get_tool/2,
       get_weeds: &DataManipulation.get_weeds/2,
       get_plants: &DataManipulation.get_plants/2,
+      get_generic_points: &DataManipulation.get_generic_points/2,
       go_to_home: &Firmware.go_to_home/2,
       grid: &DataManipulation.grid/2,
       group: &DataManipulation.group/2,
+      get_group: &DataManipulation.get_group/2,
       http: &DataManipulation.http/2,
       inspect: &DataManipulation.json_encode/2,
       json: [