+++ CHANGELOG.md
@@ -5,7 +5,8 @@
  * Fix Lua `find_home("all")` axis order bug.
  * Add `axis_order` to `move`.
  * Add `tool.seeder_tip_z_offset`.
- * Add `get_weeds` and `get_plants` Lua helpers. Add point support to `sort` Lua helper.
+ * Add `get_weeds`, `get_plants`, `get_generic_points`, and `get_group` Lua helpers.
+ * Add point support to `sort` Lua helper.
 
 # 15.4.11